@@ -1,6 +1,6 @@
 module github.com/open-policy-agent/kube-mgmt
 
-go 1.17
+go 1.22
 
 require (
 	github.com/sirupsen/logrus v1.8.1
